Added Linux 64 bit fixes for idlib.
git-svn-id: svn://0x263a.de/tekuum_src@210 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ idlib/hashing/MD4.h
@@ -10,6 +10,8 @@
 ===============================================================================
 */
 
-unsigned long MD4_BlockChecksum( const void *data, int length );
+// Techyon RB: 64 bit fix, changed long to int
+unsigned int MD4_BlockChecksum( const void *data, int length );
+// Techyon END
 
 #endif /* !__MD4_H__ */
